Move Object.prototype tests to a new file
